ARM: S5P6450: Rename LCD reset function
As there is a change of struct s3c_platform_fb which
is used to make framebuffer driver manipulate LCD
module that is specific to board schematic platform,
the name of LCD reset function is changed.

In order to use more functionality for controlling
LCD module, the more functions should be implemented.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>