FB: SAMSUNG: Modify LCD reset function
As struct s3c_platform_fb is changed which is used
to make framebuffer driver manipulate LCD module
and is specific to board, the name of LCD reset
function is changed and is possible for framebuffer
driver of S5P6450. So, the framebuffer driver should
change the name of LCD reset function for S5P6450.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>