ARM: S5PV310: Add the bootmem helper functions
This patch adds the bootmem helper functions which help
media device to know the information of memory allocation
and delineate the description for machine specific part
as to memory allocation for media device.

The description for S5PV310 machine is empty but bootmem
scheme will be implemented then this description information
will be described. The file should be in machine directory
in order to fix build error problem.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>