USB: HOST: Add INCR16/8/4 mode for S5PV210 and S5P6450
The version of USB HOST IP thatis used in S5PV210
and S5P6450 supports INCR16/8/4 mode which can do
burst transfer so the performance of USB Host will
be higher than that of USB Host which can supports
only SINGLE mode.

If you want to know as to INCR16, INCR8 and INCR4,
please refer to the AHB specification.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>