USB: HOST: Add INCR16/8/4 mode for S5PV310
The version of USB HOST IP that is used in S5PV310
supports INCR16/8/4 mode which can do burst transfer
so the performance of USB Host will be higher than
that of USB Host which can supports only SINGLE mode.

But. there is difference of bit definition between
S5PV310 and S5PV210. Therfore the define statement
of ARCH should be added.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>