USB: HOST: Add callback function for EHCI LPM
The callback function is used to enable the LPM
feature that is defined in EHCI 1.1 specification
in order to improve the power efficiency of USB
BUS used for High Speed.

LPM patch was added in commit as follows:
faf09ae5359a093161dd46c7551c5adb1b54c6bf

This patch was made by Jacob Pan who works for
Intel company, however, it is not submiited to
mainline yet.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>