FB: SAMSUNG: Modify VIDCON0 related clock selection
Register called VIDCON0 is modified since the clock scheme
of FIMD has been changed which are used for new IP version
for S5PV310.

IP clock used to be supported to LCD panel, but it is
not used anymore. So, the bit definition of VIDCON0 related
to this clock scheme and selection is reserved and should
be zero.

For more detail. refer to the datasheet for S5PV310.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>