FB: SAMSUNG: Replace physical address with virtual address as smem_start
Originaly, framebuffer interface has decided to use physical address
as the value for smem_start variable of fixed information structure
which is defined in framebuffer header file located in linux include
directory.

However, new policy is imposed that it is recommended to use virtual
address between multi-media devices instead of using physical address
as the buffer sharing in order to manipulate multi-media devices.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>