FB: SAMSUNG: Rewrite driver layer for new scheme
This patch rewrite the implementation of the samsung soc's framebuffer
driver on the layering.

The layerring is now made as follows:

          Interface layer

                 |
                 |

          Operation layer

                 |
                 |

          Hardware registers
            control layer

                 |
                 |

          Hardware registers

The layer of legacy driver is rather simple and there are a lot of
functions in one file. As multi-media devices are more complex and
have to support a lot of functionality, so, it is required that
framebuffer driver architecture be changed and rearranged.

Interface layer : This layer is in charge of infacing framebuffer
framework supported by linux kernel mainline and decides the overall
structure of driver. So, this driver is independent of the version of
hardware called FIMD.

Operation layer: This layer is in charge of operation for fimd and
take care of variables used for framebuffer framework

Hardware registers control layer: This layer is used in order to
manipulate display controller called FIMD. However, there is not
dependency of LCD panel.

The driver become more reusable and the readability is improved
than lecacy driver.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>